Move user subpages into a separate dir
