+++ physics-fe/src/router/index.js
@@ -3,7 +3,6 @@ import VueRouter from "vue-router";
 import Home from "../components/Home";
 import About from "../components/About";
 import User from "../components/User";
-import Logout from "../components/Logout";
 import TaskUpsert from "../components/TaskUpsert";
 
 Vue.use(VueRouter);
@@ -28,11 +27,6 @@ const routes = [
     path: "/user",
     name: "User",
     component: User
-  },
-  {
-    path: "/logout",
-    name: "Logout",
-    component: Logout
   }
 ];
